Clean up includes, CMakeLists, make comments more concise
