+++ sync-todo/v2/client/cpp/main.cpp
@@ -3,12 +3,12 @@
 #include <ftxui/component/loop.hpp>
 
 int main() {
-  // Initialize render destination
+  /** Initialize render destination */
   auto screen = ftxui::ScreenInteractive::FitComponent();
 
   AppController appController(&screen);
 
-  // Declare loop that handles events and renders the component to the screen
+  /** Declare loop that handles events and renders the component to the screen */
   ftxui::Loop loop(&screen, appController.component());
 
   while (!loop.HasQuitted()) {
