+++ sync-todo/v2/client/cpp/state/app_config_metadata.cpp
