+++ sync-todo/v2/client/cpp/state/app_config_metadata.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <string>
-#include <nlohmann/json.hpp>
+#include "nlohmann/json.hpp"
 
 struct AppConfigMetadata {
   std::string appId;
