+++ sync-todo/v2/client/cpp/state/app_state.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "managers/auth_manager.hpp"
-#include "managers/error_manager.hpp"
+#include "../managers/auth_manager.hpp"
+#include "../managers/error_manager.hpp"
 #include "database_state.hpp"
 
 #include "ftxui/component/screen_interactive.hpp"
