--- sync-todo/v2/client/cpp/state/home_screen_button_state.hpp
@@ -1,24 +1,0 @@
-//#pragma once
-//
-//#include "../app_state.hpp"
-//#include "ftxui/component/component.hpp"
-//
-//struct HomeScreenButtonRowState {
-//  AppState *appState;
-//
-//  std::string toggleOfflineModeButtonLabel;
-//  std::string goOfflineButtonLabel{"Go Offline"};
-//  std::string goOnlineButtonLabel{"Go Online"};
-//  ftxui::Component toggleOfflineModeButton;
-//  std::string toggleSubscriptionsButtonLabel;
-//  std::string showAllButtonLabel{"Show All Tasks"};
-//  std::string showMineButtonLabel{"Show Only My Tasks"};
-//  ftxui::Component toggleSubscriptionsButton;
-//  ftxui::Component filters;
-//  std::string logoutButtonLabel;
-//  ftxui::Component logoutButton;
-//  std::string quitButtonLabel;
-//  ftxui::Component quitButton;
-//  ftxui::Component optionsLayout;
-//  ftxui::Element optionsRowRenderer;
-//};
