+++ sync-todo/v2/client/cpp/state/offline_mode_selection.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+/** Pause the Sync connection, simulating offline network conditions.
+ *  Write while offline to see the item locally, check Atlas to see that the item does not sync. */
+enum OfflineModeSelection: int {
+  offlineModeEnabled,
+  offlineModeDisabled
+};
